Refactor WritePost and FloatingWriteButton
