+++ web/components/FeedLayout.tsx
@@ -1,12 +1,11 @@
 "use client";
 
-import useTodayWordCount from "@/hooks/useTodayWordCount";
 import { User } from "../types";
 import Feed from "./Feed";
-import FloatingWriteButton from "./FloatingWriteButton";
 import UserProfile from "./User/UserProfile";
 import WhoToFollow from "./WhoToFollow";
-import WritePostCTA from "./WritePostCTA";
+import FloatingWriteButton from "./Widgets/FloatingWriteButton";
+import WritePost from "./Widgets/WritePost";
 
 interface FeedLayoutProps {
   feedType: "fyp" | "following" | "popular";
@@ -14,15 +13,13 @@ interface FeedLayoutProps {
 }
 
 export default function FeedLayout({ feedType, user }: FeedLayoutProps) {
-  const { todayWordCount } = useTodayWordCount();
-
   return (
     <>
       <div className="w-full max-w-[1476px] mx-auto flex justify-center items-start gap-4 lg:gap-8 xl:gap-16 px-4 sm:px-6 lg:px-8 py-6">
         <div className="hidden lg:flex w-60 flex-col gap-8 flex-shrink-0 sticky top-16">
           <div className="flex flex-col gap-6">
             {user && <UserProfile user={user} />}
-            <WritePostCTA todayWordCount={todayWordCount} />
+            <WritePost />
           </div>
         </div>
